Restructured and moved FileHandling
