--- src/main/java/org/aksw/sessa/helper/files/FileHandlerInterface.java
@@ -1,17 +1,0 @@
-package org.aksw.sessa.helper.files;
-
-
-import java.io.IOException;
-import java.util.Map.Entry;
-
-/**
- * Created by Simon Bordewisch on 17.11.17.
- */
-public interface FileHandlerInterface extends AutoCloseable {
-
-  Entry<String, String> nextEntry() throws IOException;
-
-  String getFileName();
-
-  void close() throws IOException;
-}
